backend/requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.6.0
python-dotenv==1.0.1
web3==6.20.1
requests==2.32.3
cryptography==43.0.1
werkzeug==3.0.3
